Extract the different phases of OCR into sub-modules
This makes the entry point and overall dependency structure easier to grok.

In the process the `page_layout` module was removed from the public API, as it
was not intended to be exposed.
