+++ ocrs/src/recognition.rs
@@ -1,0 +1,430 @@
+use std::collections::HashMap;
+use std::error::Error;
+
+use rayon::prelude::*;
+use rten::ctc::{CtcDecoder, CtcHypothesis};
+use rten::{Dimension, FloatOperators, Model};
+use rten_imageproc::{bounding_rect, BoundingRect, Line, Point, Polygon, Rect, RotatedRect};
+use rten_tensor::prelude::*;
+use rten_tensor::{NdTensor, NdTensorView, Tensor};
+
+use crate::page_layout::line_polygon;
+use crate::preprocess::BLACK_VALUE;
+use crate::text_items::{TextChar, TextLine};
+
+// nb. The "E" before "ABCDE" should be the EUR symbol.
+const DEFAULT_ALPHABET: &str = " 0123456789!\"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~EABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
+
+/// Return the smallest multiple of `factor` that is >= `val`.
+fn round_up<
+    T: Copy
+        + std::ops::Add<T, Output = T>
+        + std::ops::Sub<T, Output = T>
+        + std::ops::Rem<T, Output = T>,
+>(
+    val: T,
+    factor: T,
+) -> T {
+    let rem = val % factor;
+    (val + factor) - rem
+}
+
+/// Details about a text line needed to prepare the input to the text
+/// recognition model.
+#[derive(Clone)]
+struct TextRecLine {
+    /// Index of this line in the list of lines found in the image.
+    index: usize,
+
+    /// Region of the image containing this line.
+    region: Polygon,
+
+    /// Width to resize this line to.
+    resized_width: u32,
+}
+
+/// Prepare an NCHW tensor containing a batch of text line images, for input
+/// into the text recognition model.
+///
+/// For each line in `lines`, the line region is extracted from `image`, resized
+/// to a fixed `output_height` and a line-specific width, then copied to the
+/// output tensor. Lines in the batch can have different widths, so the output
+/// is padded on the right side to a common width of `output_width`.
+fn prepare_text_line_batch(
+    image: &NdTensorView<f32, 3>,
+    lines: &[TextRecLine],
+    page_rect: Rect,
+    output_height: usize,
+    output_width: usize,
+) -> NdTensor<f32, 4> {
+    let mut output = NdTensor::zeros([lines.len(), 1, output_height, output_width]);
+    output.apply(|_| BLACK_VALUE);
+
+    // Page rect adjusted to only contain coordinates that are valid for
+    // indexing into the input image.
+    let page_index_rect = page_rect.adjust_tlbr(0, 0, -1, -1);
+
+    for (group_line_index, line) in lines.iter().enumerate() {
+        let grey_chan = image.slice([0]);
+
+        let line_rect = line.region.bounding_rect();
+        let mut line_img =
+            NdTensor::zeros([line_rect.height() as usize, line_rect.width() as usize]);
+        line_img.apply(|_| BLACK_VALUE);
+
+        for in_p in line.region.fill_iter() {
+            let out_p = Point::from_yx(in_p.y - line_rect.top(), in_p.x - line_rect.left());
+            if !page_index_rect.contains_point(in_p) || !page_index_rect.contains_point(out_p) {
+                continue;
+            }
+            line_img[[out_p.y as usize, out_p.x as usize]] =
+                grey_chan[[in_p.y as usize, in_p.x as usize]];
+        }
+
+        let resized_line_img = line_img
+            .reshaped([1, 1, line_img.size(0), line_img.size(1)])
+            .resize_image([output_height, line.resized_width as usize])
+            .unwrap();
+
+        let resized_line_img: NdTensorView<f32, 2> =
+            resized_line_img.squeezed().try_into().unwrap();
+
+        output
+            .slice_mut((group_line_index, 0, .., ..(line.resized_width as usize)))
+            .copy_from(&resized_line_img);
+    }
+
+    output
+}
+
+/// Return the bounding rectangle of the slice of a polygon with X coordinates
+/// between `min_x` and `max_x` inclusive.
+fn polygon_slice_bounding_rect(
+    poly: Polygon<i32, &[Point]>,
+    min_x: i32,
+    max_x: i32,
+) -> Option<Rect> {
+    poly.edges()
+        .filter_map(|e| {
+            let e = e.rightwards();
+
+            // Filter out edges that don't overlap [min_x, max_x].
+            if (e.start.x < min_x && e.end.x < min_x) || (e.start.x > max_x && e.end.x > max_x) {
+                return None;
+            }
+
+            // Truncate edge to [min_x, max_x].
+            let trunc_edge_start = e
+                .to_f32()
+                .y_for_x(min_x as f32)
+                .map(|y| Point::from_yx(y.round() as i32, min_x))
+                .unwrap_or(e.start);
+
+            let trunc_edge_end = e
+                .to_f32()
+                .y_for_x(max_x as f32)
+                .map(|y| Point::from_yx(y.round() as i32, max_x))
+                .unwrap_or(e.end);
+
+            Some(Line::from_endpoints(trunc_edge_start, trunc_edge_end))
+        })
+        .fold(None, |bounding_rect, e| {
+            let edge_br = e.bounding_rect();
+            bounding_rect.map(|br| br.union(edge_br)).or(Some(edge_br))
+        })
+}
+
+/// Method used to decode sequence model outputs to a sequence of labels.
+///
+/// See [CtcDecoder] for more details.
+#[derive(Copy, Clone, Default)]
+pub enum DecodeMethod {
+    #[default]
+    Greedy,
+    BeamSearch {
+        width: u32,
+    },
+}
+
+#[derive(Clone, Default)]
+pub struct RecognitionOpt {
+    pub debug: bool,
+
+    /// Method used to decode character sequence outputs to character values.
+    pub decode_method: DecodeMethod,
+}
+
+/// Input and output from recognition for a single text line.
+struct LineRecResult {
+    /// Input to the recognition model.
+    line: TextRecLine,
+
+    /// Length of input sequences to recognition model, padded so that all
+    /// lines in batch have the same length.
+    rec_input_len: usize,
+
+    /// Length of output sequences from recognition model, used as input to
+    /// CTC decoding.
+    ctc_input_len: usize,
+
+    /// Output label sequence produced by CTC decoding.
+    ctc_output: CtcHypothesis,
+}
+
+/// Combine information from the input and output of text line recognition
+/// to produce [TextLine]s containing character sequences and bounding boxes
+/// for each line.
+///
+/// Entries in the result may be `None` if no text was recognized for a line.
+fn text_lines_from_recognition_results(results: &[LineRecResult]) -> Vec<Option<TextLine>> {
+    results
+        .iter()
+        .map(|result| {
+            let line_rect = result.line.region.bounding_rect();
+            let x_scale_factor = (line_rect.width() as f32) / (result.line.resized_width as f32);
+
+            // Calculate how much the recognition model downscales the image
+            // width. We assume this will be an integer factor, or close to it
+            // if the input width is not an exact multiple of the downscaling
+            // factor.
+            let downsample_factor =
+                (result.rec_input_len as f32 / result.ctc_input_len as f32).round() as u32;
+
+            let steps = result.ctc_output.steps();
+            let text_line: Vec<TextChar> = steps
+                .iter()
+                .enumerate()
+                .filter_map(|(i, step)| {
+                    // X coord range of character in line recognition input image.
+                    let start_x = step.pos * downsample_factor;
+                    let end_x = if let Some(next_step) = steps.get(i + 1) {
+                        next_step.pos * downsample_factor
+                    } else {
+                        result.line.resized_width
+                    };
+
+                    // Map X coords to those of the input image.
+                    let [start_x, end_x] = [start_x, end_x]
+                        .map(|x| line_rect.left() + (x as f32 * x_scale_factor) as i32);
+
+                    // Since the recognition input is padded, it is possible to
+                    // get predicted characters in the output with positions
+                    // that correspond to the padding region, and thus are
+                    // outside the bounds of the original line. Ignore these.
+                    if start_x >= line_rect.right() {
+                        return None;
+                    }
+
+                    let char = DEFAULT_ALPHABET
+                        .chars()
+                        .nth((step.label - 1) as usize)
+                        .unwrap_or('?');
+
+                    Some(TextChar {
+                        char,
+                        rect: polygon_slice_bounding_rect(
+                            result.line.region.borrow(),
+                            start_x,
+                            end_x,
+                        )
+                        .expect("invalid X coords"),
+                    })
+                })
+                .collect();
+
+            if text_line.is_empty() {
+                None
+            } else {
+                Some(TextLine::new(text_line))
+            }
+        })
+        .collect()
+}
+
+/// Encapsulates validation and execution of the text line recognition model.
+pub struct RecognitionModel {
+    model: Model,
+    input_id: usize,
+    input_shape: Vec<Dimension>,
+    output_id: usize,
+}
+
+impl RecognitionModel {
+    /// Validate that a model has the expected inputs and outputs for
+    /// a text recognition model and wrap it as a [RecognitionModel].
+    pub fn from_model(model: Model) -> Result<RecognitionModel, Box<dyn Error>> {
+        let input_id = model
+            .input_ids()
+            .first()
+            .copied()
+            .ok_or("recognition model has no inputs")?;
+        let input_shape = model
+            .node_info(input_id)
+            .and_then(|info| info.shape())
+            .ok_or("recognition model does not specify input shape")?;
+        let output_id = model
+            .output_ids()
+            .first()
+            .copied()
+            .ok_or("recognition model has no outputs")?;
+        Ok(RecognitionModel {
+            model,
+            input_id,
+            input_shape: input_shape.into_iter().collect(),
+            output_id,
+        })
+    }
+
+    /// Return the expected height of input line images.
+    fn input_height(&self) -> u32 {
+        match self.input_shape[2] {
+            Dimension::Fixed(size) => size.try_into().unwrap(),
+            Dimension::Symbolic(_) => 50,
+        }
+    }
+
+    /// Run text recognition on an NCHW batch of text line images, and return
+    /// a `[batch, seq, label]` tensor of class probabilities.
+    fn run(&self, input: NdTensor<f32, 4>) -> Result<NdTensor<f32, 3>, Box<dyn Error>> {
+        let input: Tensor<f32> = input.into();
+        let [output] =
+            self.model
+                .run_n(&[(self.input_id, (&input).into())], [self.output_id], None)?;
+        let mut rec_sequence: NdTensor<f32, 3> = output.try_into()?;
+
+        // Transpose from [seq, batch, class] => [batch, seq, class]
+        rec_sequence.permute([1, 0, 2]);
+
+        Ok(rec_sequence)
+    }
+}
+
+/// Recognize text lines in an image.
+///
+/// `image` is a CHW greyscale image with values in the range `ZERO_VALUE` to
+/// `ZERO_VALUE + 1`. `lines` is a list of detected text lines, where each line
+/// is a sequence of word rects. `model` is a recognition model which accepts an
+/// NCHW tensor of greyscale line images and outputs a `[sequence, batch, label]`
+/// tensor of log probabilities of character classes, which must be converted to
+/// a character sequence using CTC decoding.
+///
+/// Entries in the result can be `None` if no text was found in a line.
+pub fn recognize_text_lines(
+    image: NdTensorView<f32, 3>,
+    lines: &[Vec<RotatedRect>],
+    model: &RecognitionModel,
+    opts: RecognitionOpt,
+) -> Result<Vec<Option<TextLine>>, Box<dyn Error>> {
+    let RecognitionOpt {
+        debug,
+        decode_method,
+    } = opts;
+
+    let [_, img_height, img_width] = image.shape();
+    let page_rect = Rect::from_hw(img_height as i32, img_width as i32);
+
+    // Compute width to resize a text line image to, for a given height.
+    fn resized_line_width(orig_width: i32, orig_height: i32, height: i32) -> u32 {
+        // Min/max widths for resized line images. These must match the PyTorch
+        // `HierTextRecognition` dataset loader.
+        let min_width = 10.;
+        let max_width = 800.;
+        let aspect_ratio = orig_width as f32 / orig_height as f32;
+        (height as f32 * aspect_ratio).max(min_width).min(max_width) as u32
+    }
+
+    // Group lines into batches which will have similar widths after resizing
+    // to a fixed height.
+    //
+    // It is more efficient to run recognition on multiple lines at once, but
+    // all line images in a batch must be padded to an equal length. Some
+    // computation is wasted on shorter lines in the batch. Choosing batches
+    // such that all line images have a similar width reduces this wastage.
+    // There is a trade-off between maximizing the batch size and minimizing
+    // the variance in width of images in the batch.
+    let rec_img_height = model.input_height();
+    let mut line_groups: HashMap<i32, Vec<TextRecLine>> = HashMap::new();
+    for (line_index, word_rects) in lines.iter().enumerate() {
+        let line_rect = bounding_rect(word_rects.iter())
+            .expect("line has no words")
+            .integral_bounding_rect();
+        let resized_width =
+            resized_line_width(line_rect.width(), line_rect.height(), rec_img_height as i32);
+        let group_width = round_up(resized_width, 50);
+        line_groups
+            .entry(group_width as i32)
+            .or_default()
+            .push(TextRecLine {
+                index: line_index,
+                region: Polygon::new(line_polygon(word_rects)),
+                resized_width,
+            });
+    }
+
+    // Split large line groups up into smaller batches that can be processed
+    // in parallel.
+    let max_lines_per_group = 20;
+    let line_groups: Vec<(i32, Vec<TextRecLine>)> = line_groups
+        .into_iter()
+        .flat_map(|(group_width, lines)| {
+            lines
+                .chunks(max_lines_per_group)
+                .map(|chunk| (group_width, chunk.to_vec()))
+                .collect::<Vec<_>>()
+        })
+        .collect();
+
+    // Run text recognition on batches of lines.
+    let mut line_rec_results: Vec<LineRecResult> = line_groups
+        .into_par_iter()
+        .flat_map(|(group_width, lines)| {
+            if debug {
+                println!(
+                    "Processing group of {} lines of width {}",
+                    lines.len(),
+                    group_width,
+                );
+            }
+
+            let rec_input = prepare_text_line_batch(
+                &image,
+                &lines,
+                page_rect,
+                rec_img_height as usize,
+                group_width as usize,
+            );
+
+            // TODO - Propagate errors from recognition model to caller.
+            let rec_output = model.run(rec_input).expect("recognition failed");
+            let ctc_input_len = rec_output.shape()[1];
+
+            // Apply CTC decoding to get the label sequence for each line.
+            lines
+                .into_iter()
+                .enumerate()
+                .map(|(group_line_index, line)| {
+                    let decoder = CtcDecoder::new();
+                    let input_seq = rec_output.slice([group_line_index]);
+                    let ctc_output = match decode_method {
+                        DecodeMethod::Greedy => decoder.decode_greedy(input_seq),
+                        DecodeMethod::BeamSearch { width } => decoder.decode_beam(input_seq, width),
+                    };
+                    LineRecResult {
+                        line,
+                        rec_input_len: group_width as usize,
+                        ctc_input_len,
+                        ctc_output,
+                    }
+                })
+                .collect::<Vec<_>>()
+        })
+        .collect();
+
+    // The recognition outputs are in a different order than the inputs due to
+    // batching and parallel processing. Re-sort them into input order.
+    line_rec_results.sort_by_key(|result| result.line.index);
+
+    let text_lines = text_lines_from_recognition_results(&line_rec_results);
+
+    Ok(text_lines)
+}
